requirements.txt
# Core GPU Libraries
cupy-cuda12x==13.3.0
cudf-cu12==25.2.0
cuml-cu12==25.2.0
rapids-cu12>=25.2.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Model Persistence
joblib>=1.3.0

# Python Environment
python>=3.8.0